Running tasklist now runs menu
Cool app!

Changed __main__ to run the menu.
Removed the unnecessary check in `_renumber_tasks()`.
What editor are you using?  It puts invisible characters on each blank line.
Why renumber the tasks?  Is it merely for display purposes or is there a deeper reason?
Should TaskList be a dict instead of a list?